Restyle the real Setup (register) wizard to the design system — Phase 1
Restyle the create-account wizard (UI only; onboarding state machine, lanes,
validation, handlers, and useSessionFocus markers preserved). Extracted the
stage sub-components into setup/ and restyled each:

- Shell wrapped in AuthLayout (galaxy); actions row → Button (submit via
  form.requestSubmit; data-setup-submit kept).
- Welcome mode cards → Surface (data-setup-lane preserved).
- Account fields → TextInput (+ inputProps autoComplete).
- Preferences core → Checkbox + TextInput + Select (admin, timezone).
- Review summary → Surface cards (data-setup-summary-* kept).
- Sidebar rail → Stepper; heading → type utilities.
- Button/Surface gained a dataAttrs passthrough for focus markers.

Deferred to Phase 2 (left old-styled): the advanced Custom-lane sections
(AI / system files / device) and the GuidePanel chat assistant.
