+++ web/src/app/features/session/setup/AccountDetails.css
@@ -1,0 +1,34 @@
+/* ============================================================
+   AccountDetails — create-account "Account" step, restyled to
+   the GSV design system. Stacked TextInput fields + a password
+   safety note, using the periwinkle token vocabulary.
+   ============================================================ */
+
+.account-details-fields {
+  display: flex;
+  flex-direction: column;
+  gap: 18px;
+}
+
+.account-details-note {
+  margin-top: 22px;
+  padding: 14px 16px;
+  border: 1px solid var(--border);
+  background: var(--panel-2);
+}
+.account-details-note strong {
+  display: block;
+  font-family: var(--gsv-font-mono);
+  font-size: 9.5px;
+  letter-spacing: 0.22em;
+  text-transform: uppercase;
+  color: var(--accent);
+  margin-bottom: 6px;
+}
+.account-details-note p {
+  margin: 0;
+  font-family: var(--gsv-font-prose);
+  font-size: 12px;
+  line-height: 1.5;
+  color: var(--text-dim);
+}
